requirements.txt
altgraph==0.17.4
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
customtkinter==5.2.2
darkdetect==0.8.0
et_xmlfile==2.0.0
openai>=1.0.0
openpyxl==3.1.5
packaging==25.0
pdfminer.six==20250327
pdfplumber==0.11.6
pefile==2023.2.7
pillow==11.2.1
pycparser==2.22
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
PyMuPDF==1.26.0
pypdfium2==4.30.1
pywin32-ctypes==0.2.3
python-dotenv>=1.0.0
setuptools==80.9.0
PyQt5>=5.15
pandas>=2.0
jsonschema>=4.0
